+++ Model/Contas.cs
@@ -10,10 +10,10 @@ namespace Controle_Financeiro_Pessoal.Model
     {
         public int Id_Conta { get; set; }
         public string Conta { get; set; }
-        public int Saldo { get; set; }
+        public double Saldo { get; set; }
 
 
-        public Contas(int idConta, string conta, int saldo)
+        public Contas(int idConta, string conta, double saldo)
         {
             Id_Conta = idConta;
             Conta = conta;
